requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.2
django-cors-headers==4.2.0
Pillow==10.0.0
python-decouple==3.8
psycopg2-binary==2.9.7
drf-yasg==1.21.7
gunicorn==21.2.0
